refactor: CustomView 파일로 분리하여 폴더로 묶음, 주석 처리
